split components in two header and shops
