+++ src/App.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react'
 
 import { AppProvider, AppConsumer } from './context'
-import Header from './components/Header'
-import Tabs from './components/Tabs'
+import Header from './components/Header/Header'
+import Tabs from './components/Shops/Tabs'
 
 class App extends Component {
   render () {
