+++ src/components/Shops/TabContainer.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import List from '@material-ui/core/List'
 
-import { AppConsumer } from '../context'
+import { AppConsumer } from '../../context'
 import Product from './Product'
 
 class TabContainer extends Component {
